Version 0.8
Updated to a more reliable webserver instead of a proxy, much safer as
well since the only data transfered is a 32 byte cookie. The cookie is
also renewed on each page load on the crunchyroll website at most once a
minute, as well as once when the browser first renews a session.